shm_open: unlink bad shared memory
Similar to state saved in files, if the shared memory segment is bad
(less than the size of the state struct), delete it and exit 127+EFAULT.

For state saved in files, also exit. Previously the state was silently
removed and recreated.

Also modify the seccomp policy to allow deleting state.