runlimit: exec command
Re-write runlimit to exec(3) a command if the rate is below the
threshold. runlimit will exit with status 111 if the threshold is
reached or can optionally wait for the next run interval to exec the
command.

Remove support for sandboxing since the exec()'ed process would inherit
the restrictions.

Remove support for using shared memory for storing the rate state
because it simpler and more intuitive to use a file for state.